+++ grub-reboot-picker/grub-reboot-picker.py
@@ -2,23 +2,25 @@
 import gi
 import os
 import re
+import subprocess
 gi.require_version("Gtk", "3.0")
-gi.require_version('AppIndicator3', '0.1')
-from gi.repository import Gtk, AppIndicator3
+from gi.repository import Gtk
+
+try:
+    gi.require_version('AyatanaAppIndicator3', '0.1')
+    from gi.repository import AyatanaAppIndicator3 as AppIndicator3
+except (ValueError, ImportError):
+    gi.require_version('AppIndicator3', '0.1')
+    from gi.repository import AppIndicator3
 
 
 SHOW_GRUB_MENU_SUB_MENUS = True
-DEVELOPMENT_MODE = False
-GRUB_CONFIG_PATH = "/boot/grub/grub.cfg"
-if DEVELOPMENT_MODE:
-    GRUB_CONFIG_PATH = os.path.join(os.path.dirname(os.path.realpath(__file__)),  "grub.test3.cfg")
-    GRUB_CONFIG_PATH = "/boot/grub/grub.cfg"
-
-icon_name = "un-reboot"
+DEVELOPMENT_MODE = os.environ.get("DEBUG", False)
 
 
-def get_all_grub_entries(file_path, include_submenus=True):
+def get_all_grub_entries(include_submenus=True):
     """
+    Runs grub-mkconfig and gets the grub output.
     Build a dictionary of Grub menu items with sub menu items if applicable.
     Simply if it has child items it's a 'submenu' else it's just a top level menu.
     {
@@ -32,12 +34,13 @@ def get_all_grub_entries(file_path, include_submenus=True):
         'UEFI Firmware Settings': []
     }
     """
-    with open(file_path, 'r') as file:
-        lines = file.readlines()
 
-    menu_pattern = re.compile("^\\s*menuentry ['\"]([^'\"]*)['\"]")
-    submenu_pattern = re.compile("^\\s*submenu ['\"]([^']*)['\"]")
-    closing_brace_pattern = re.compile("^\\s*}")
+    grub_mkconfig_output = subprocess.check_output(["pkexec", "grub-mkconfig"], stderr=subprocess.STDOUT)
+    lines = grub_mkconfig_output.decode("utf-8").splitlines()
+
+    menu_pattern = re.compile(r"^\s*menuentry ['\"]([^'\"]*)['\"]")
+    submenu_pattern = re.compile(r"^\s*submenu ['\"]([^'\"]*)['\"]")
+    closing_brace_pattern = re.compile(r"^\s*}")
 
     menu_entries = {}
 
@@ -76,26 +79,95 @@ def get_all_grub_entries(file_path, include_submenus=True):
     return menu_entries
 
 
+def get_grub_entries_with_args(grub_entries):
+    """
+    Returns a list of Grub menu items with the arguments for grub-reboot.
+    [
+        {
+        "title": "Ubuntu",
+        "grub_reboot_args": "Ubuntu"
+        },
+        {
+            "title": "Advanced options for Ubuntu",
+            "children": [
+                {
+                "title": "Ubuntu, with Linux 6.8.0-39-generic",
+                "grub_reboot_args": "Advanced options for Ubuntu>Ubuntu, with Linux 6.8.0-39-generic",
+                },
+                {
+                "title": "Ubuntu, with Linux 6.8.0-39-generic (recovery mode)",
+                "grub_reboot_args": "Advanced options for Ubuntu>Ubuntu, with Linux 6.8.0-39-generic (recovery mode)",
+        },
+        {
+            "title": "Memory test (memtest86+x64.bin)",
+            "grub_reboot_args": "Memory test (memtest86+x64.bin)",
+        },
+        {
+            "title": "Memory test (memtest86+x64.bin, serial console)",
+            "grub_reboot_args": "Memory test (memtest86+x64.bin, serial console)",
+        },
+        {
+            "title": "UEFI Firmware Settings",
+            "grub_reboot_args": "UEFI Firmware Settings",
+    ]
+    """
+
+    grub_entries_with_args = []
+    for title, children in grub_entries:
+        if len(children) > 0:
+            child_entries = []
+            for child in children:
+                child_entries.append({
+                    "title": child,
+                    "grub_reboot_args": f"{title}>{child}",
+                })
+            grub_entries_with_args.append({
+                "title": title,
+                "children": child_entries
+            })
+        else:
+            grub_entries_with_args.append({
+                "title": title,
+                "grub_reboot_args": title,
+            })
+
+    return grub_entries_with_args
+
+
 def build_menu():
     menu = Gtk.Menu()
 
-    grub_entries = get_all_grub_entries(GRUB_CONFIG_PATH, SHOW_GRUB_MENU_SUB_MENUS)
+    grub_entries = get_all_grub_entries(SHOW_GRUB_MENU_SUB_MENUS)
+    grub_entries_with_args = get_grub_entries_with_args(grub_entries.items())
 
-    print(grub_entries)
+    print(grub_entries_with_args)
 
-    for grub_entry, grub_children in grub_entries.items():
-        menuitem = Gtk.MenuItem(label=grub_entry)
-        if len(grub_children) == 0:
-            menuitem.connect('activate', do_grub_reboot, grub_entry)
+    menu_item_memory_test = Gtk.MenuItem(label="Memory Test")
+    sub_menu_memory_test = Gtk.Menu()
+
+    for entries in grub_entries_with_args:
+
+        if "children" in entries and len(entries["children"]) > 0:
+            menu_item = Gtk.MenuItem(label=entries["title"])
+            sub_menu = Gtk.Menu()
+            for child in entries["children"]:
+                sub_menu_item = Gtk.MenuItem(label=child["title"])
+                sub_menu_item.connect('activate', do_grub_reboot, child["grub_reboot_args"])
+                sub_menu.append(sub_menu_item)
+            menu_item.set_submenu(sub_menu)
+            menu.append(menu_item)
+        elif entries["title"].startswith("Memory test"):
+            submenu_item_memory_test = Gtk.MenuItem(label=entries["title"])
+            submenu_item_memory_test.connect('activate', do_grub_reboot, entries["grub_reboot_args"])
+            sub_menu_memory_test.append(submenu_item_memory_test)
         else:
-            submenu = Gtk.Menu()
-            for grub_child in grub_children:
-                submenu_item = Gtk.MenuItem(label=grub_child)
-                submenu_item.connect('activate', do_grub_reboot, grub_child, grub_entry)
-                submenu.append(submenu_item)
-            menuitem.set_submenu(submenu)
+            menu_item = Gtk.MenuItem(label=entries["title"])
+            menu_item.connect('activate', do_grub_reboot, entries["grub_reboot_args"])
+            menu.append(menu_item)
 
-        menu.append(menuitem)
+    if len(sub_menu_memory_test.get_children()) > 0:
+        menu_item_memory_test.set_submenu(sub_menu_memory_test)
+        menu.append(menu_item_memory_test)
 
     shutdown_item = Gtk.MenuItem(label='Shutdown')
     shutdown_item.connect('activate', do_shutdown)
@@ -108,6 +180,7 @@ def build_menu():
     menu.show_all()
     return menu
 
+
 # Returns the correct version of the given command, depending on whether
 # molly-guard is installed.
 def molly_command(command):
@@ -118,24 +191,19 @@ def molly_command(command):
         # Maybe this should return "/sbin/command"
         return command
 
-def do_grub_reboot(menuitem, grub_entry, parent_grub_entry=None):
-    if parent_grub_entry is not None:
-        grub_reboot_value = "{}>{}".format(parent_grub_entry, grub_entry)
-    else:
-        grub_reboot_value = "{}".format(grub_entry)
 
+def do_grub_reboot(_, grub_reboot_args):
     reboot_command = molly_command("reboot")
 
-    if DEVELOPMENT_MODE:
-        print("pkexec grub-reboot '{}' && sleep 1 && pkexec {}".format(grub_reboot_value, reboot_command))
+    print("pkexec grub-reboot '{}' && sleep 1 && pkexec {}".format(grub_reboot_args, reboot_command))
     if not DEVELOPMENT_MODE:
-        os.system("pkexec grub-reboot '{}' && sleep 1 && pkexec {}".format(grub_reboot_value, reboot_command))
+        os.system("pkexec grub-reboot '{}' && sleep 1 && pkexec {}".format(grub_reboot_args, reboot_command))
+
 
 def do_shutdown(_):
     shutdown_command = molly_command("shutdown")
 
-    if DEVELOPMENT_MODE:
-        print("sleep 1 && pkexec {} -h now".format(shutdown_command))
+    print("sleep 1 && pkexec {} -h now".format(shutdown_command))
     if not DEVELOPMENT_MODE:
         os.system("sleep 1 && pkexec {} -h now".format(shutdown_command))
 
@@ -143,6 +211,14 @@ def do_shutdown(_):
 def quit(_):
     Gtk.main_quit()
 
+
+# The icon ought to get deployed to /usr/share/icons/hicolor/scalable/apps/,
+# so it can just be referenced by name.
+icon_name = "grub-reboot-picker"
+
+# Local development
+if os.path.exists("./assets/grub-reboot-picker.svg"):
+    icon_name = os.path.abspath("./assets/grub-reboot-picker.svg")
 
 indicator = AppIndicator3.Indicator.new(
     "customtray", icon_name,